etl/extract/requirements.txt
websocket-client==1.8.0
finnhub-python==2.4.23